Fix integer ABI and protocol parameters in stage1
Preserve target-sized Elisa integers while correcting C ABI declarations and wasm bindings. Lower protocol value parameters to constrained generics, improve matching diagnostics, and add runtime parity coverage.

Retain declaration/reference metadata for tooling and prevent metadata from interfering with effects and exports. Include parser grant fixes, documentation, and the semantic oracle update.

+++ scripts/wasm_export_scan.py
@@ -23,7 +23,7 @@ SCALAR_TYPES = {
     "u32": "i32",
     "i64": "i64",
     "u64": "i64",
-    "int": "i64",
+    "int": "i32",
     "char": "i64",
     "isize": "i32",
     "usize": "i32",
